+++ example_test.go
@@ -88,6 +88,7 @@ func ExampleIndex_Local() {
 	local := index.Local()
 	var matches []string
 	for _, region := range []int{20, 30} {
+		matches = matches[:0]
 		local.Search(constraint{store: pointer(10), region: pointer(region)}, &matches)
 		fmt.Println(matches)
 	}